core/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==4.2.17
django-ckeditor==6.7.2
django-js-asset==3.0.1
Faker==33.1.0
idna==3.10
pillow==11.0.0
psycopg==3.2.3
psycopg-binary==3.2.3
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
